# billingwebapp/docs/openapi-ai-v1.yaml
openapi: 3.0.3
info:
  title: Clinic AI Integration API
  version: 1.0.0
  description: Secure server-to-server API for the separate Call Assistant. Direct database access is unsupported.
servers:
  - url: /api/v1/ai
security:
  - bearerAuth: []
paths:
...
  /clinic/status:
    get:
      summary: Effective live status for one or all active locations
      x-required-scope: clinic:read
      parameters: [{name: location_id, in: query, schema: {type: integer, minimum: 1}}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
  /locations/{location_id}/share:
    get:
      summary: Public location and map information
      x-required-scope: clinic:read
      parameters: [{$ref: '#/components/parameters/LocationId'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /locations/{location_id}/send:
    post:
      summary: Send approved CLINIC_LOCATION template to the verified patient
      x-required-scope: notification:send
      parameters: [{$ref: '#/components/parameters/LocationId'}, {$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/Channel'}
      responses: {'202': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}
  /doctors:
    get:
      summary: Active public doctors only
      x-required-scope: doctor:read
      parameters:
        - {name: specialty, in: query, schema: {type: string, maxLength: 120}}
        - {name: location_id, in: query, schema: {type: integer, minimum: 1}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /doctors/{doctor_id}/schedule:
    get:
      summary: Effective clinic plus doctor schedule
      x-required-scope: doctor:read
      parameters: [{$ref: '#/components/parameters/DoctorId'}, {$ref: '#/components/parameters/LocationQuery'}, {$ref: '#/components/parameters/DateQuery'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /doctors/{doctor_id}/availability:
    get:
      summary: Effective availability and next available date
      x-required-scope: doctor:read
      parameters: [{$ref: '#/components/parameters/DoctorId'}, {$ref: '#/components/parameters/LocationQuery'}, {$ref: '#/components/parameters/DateQuery'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /knowledge:
    get:
      summary: Approved active non-clinical knowledge
      x-required-scope: clinic:read
      parameters:
        - {name: category, in: query, schema: {type: string, maxLength: 80}}
        - {name: language, in: query, schema: {type: string, maxLength: 12}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /reception/status:
    get:
      summary: Live reception status, next opening and approved transfer destination
      x-required-scope: clinic:read
      parameters: [{$ref: '#/components/parameters/LocationQuery'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
...
  /callbacks:
    post:
      summary: Create idempotent callback request
      x-required-scope: callback:create
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {$ref: '#/components/responses/Success'}}
  /complaints:
    post:
      summary: Create idempotent complaint record
      x-required-scope: complaint:create
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {$ref: '#/components/responses/Success'}}
  /notifications:
    post:
      summary: Send an approved template only to verified patient number
      x-required-scope: notification:send
      parameters: [{$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'202': {$ref: '#/components/responses/Success'}}
  /notifications/appointment:
    post:
      summary: Send approved APPOINTMENT_CONFIRMATION for owned appointment
      x-required-scope: notification:send
      parameters: [{$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'202': {$ref: '#/components/responses/Success'}}
components:
  securitySchemes:
    basicAuth: {type: http, scheme: basic}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
  parameters:
    DoctorId: {name: doctor_id, in: path, required: true, schema: {type: integer, minimum: 1}}
    LocationId: {name: location_id, in: path, required: true, schema: {type: integer, minimum: 1}}
    LocationQuery: {name: location_id, in: query, required: true, schema: {type: integer, minimum: 1}}
    DateQuery: {name: date, in: query, required: true, schema: {type: string, format: date}}
...
    PatientSession: {name: X-Patient-Session, in: header, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
  requestBodies:
    Channel:
      required: true
      content: {application/json: {schema: {type: object, properties: {channel: {type: string, enum: [SMS, WHATSAPP, EMAIL]}, language: {type: string, default: en}}}}}
  schemas:
...
    Success:
      type: object
      required: [success, data, request_id]
      properties: {success: {type: boolean, enum: [true]}, data: {type: object}, request_id: {type: string}}
    Error:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, enum: [false]}
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
        request_id: {type: string}
  responses:
    Success: {description: Standard success envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Success'}}}}
...
    Conflict: {description: State, capacity or idempotency conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...